+++ README.md
@@ -1,1 +1,31 @@
 # react-gift-order
+
+## 로그인 기능 요구사항
+
+### 🔸 ID (이메일)
+
+- [x] ID는 반드시 입력 되어야 해요.
+- [x] ID는 이메일 형식이어야 해요.
+- [x] 만약 ID Input Field를 벗어났을 때 위 조건이 충족되지 않으면 에러 메시지를 보여줘요. (ex. ID를 입력해주세요., ID는 이메일 형식으로 입력해주세요.)
+- [x] 위 조건이 충족되면 에러 메시지가 사라져요."
+
+---
+
+### 🔸 PW (비밀번호)
+
+- [x] PW는 최소 8글자 이상 입력되어야 해요.
+- [x] 만약 PW Input Field를 벗어났을 때 위 조건이 충족되지 않으면 에러 메시지를 보여줘요. (ex. PW를 입력해주세요. PW는 최소 8글자 이상이어야 합니다.)
+
+---
+
+### 🔸 로그인 버튼
+
+- [x] ID의 조건과 PW 조건이 모두 충족 되었을 때에만 버튼이 활성화 돼요.
+
+---
+
+## 🔁 리팩터링
+
+- [x] 만약 위 로직을 구현하면서 커스텀 훅을 사용하지 않았다면 재사용을 고려해 커스텀 훅으로 로그인 기능을 리팩터링하세요. (LoginFormSection 에 직접적으로 useState를 사용하지 않는 방향으로 만들어보세요.)
+
+---
